up hash algorithm and image library

+++ src/render/gl/glsl/triangles.glsl
@@ -42,7 +42,7 @@ void main() {
 #endif
 
 #ifdef Qk_SHADER_IF_FLAGS_AACLIP
-	float alpha = smoothstep(0.9, 1.0, texelFetch(aaclip, ivec2(gl_FragCoord.xy), 0).r);
+	lowp float alpha = smoothstep(0.9, 1.0, texelFetch(aaclip, ivec2(gl_FragCoord.xy), 0).r);
 	// apply premultiplied alpha
 	// float premul = mix(1.0, alpha, premultipliedAlpha);
 	// fragColor *= vec4(vec3(premul), alpha);
